requirements.txt
amqp==5.3.1
asgiref==3.8.1
asttokens==3.0.0
attrs==25.3.0
billiard==4.2.1
celery==5.4.0
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cron-descriptor==1.4.5
decorator==5.2.1
Django==5.1.7
django-appconf==1.1.0
django-celery-beat==2.7.0
django-debug-toolbar==5.1.0
django-filter==25.1
django-imagekit==5.0.0
django-js-asset==3.1.2
django-mptt==0.16.0
django-ratelimit==4.1.0
django-redis==5.4.0
django-taggit==6.1.0
django-timezone-field==7.1
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
dotenv==0.9.9
drf-spectacular==0.28.0
executing==2.2.0
inflection==0.5.1
ipython==9.0.2
ipython_pygments_lexers==1.1.1
jedi==0.19.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
kombu==5.5.1
matplotlib-inline==0.1.7
parso==0.8.4
pexpect==4.9.0
pilkit==3.0
pillow==11.1.0
prompt_toolkit==3.0.50
psycopg2-binary==2.9.10
ptyprocess==0.7.0
pure_eval==0.2.3
Pygments==2.19.1
PyJWT==2.9.0
pyotp==2.9.0
python-crontab==3.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
PyYAML==6.0.2
redis==5.2.1
referencing==0.36.2
rpds-py==0.24.0
six==1.17.0
sqlparse==0.5.3
stack-data==0.6.3
traitlets==5.14.3
tzdata==2025.1
uritemplate==4.1.1
vine==5.1.0
wcwidth==0.2.13
